# Grp2BackendAN/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "thinkbridge.Grp2BackendAN.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/ToDoMaster/GetAllToDoMaster": {
      "post": {
        "tags": [
          "ToDoMaster"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoMasterReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoMasterReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoMasterReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoMasterReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoMaster/GetToDoMasterById/{ToDoTaskId}/{withDetails}": {
      "get": {
        "tags": [
          "ToDoMaster"
        ],
        "parameters": [
          {
            "name": "ToDoTaskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "withDetails",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDetailDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDetailDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDetailDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoMaster/AddToDoMaster": {
      "post": {
        "tags": [
          "ToDoMaster"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoMasterReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoMasterReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoMasterReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoMasterReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoMaster/UpdateToDoMaster": {
      "post": {
        "tags": [
          "ToDoMaster"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoMasterReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoMasterReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoMasterReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoMasterReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoMasterResDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoMaster/DeleteToDoMaster/{ToDoTaskId}": {
      "delete": {
        "tags": [
          "ToDoMaster"
        ],
        "parameters": [
          {
            "name": "ToDoTaskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddToDoMasterReqDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "repeatFrequency": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "isAssignedToTeam": {
            "type": "boolean"
          },
          "hasChecklist": {
            "type": "boolean"
          },
          "hasReminder": {
            "type": "boolean"
          },
          "percentageCompleted": {
            "type": "number",
            "format": "double"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "mediaAttachmentType": {
            "type": "string",
            "nullable": true
          },
          "mediaAttachmentURL": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BooleanFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BooleanFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BooleanFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "boolean"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "ComparisonOperator": {
        "enum": [
          "Equals",
          "NotEquals",
          "GreaterThan",
          "LessThan",
          "GreaterThanOrEqual",
          "LessThanOrEqual",
          "Contains"
        ],
        "type": "string"
      },
      "DateTimeFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateTimeFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateTimeFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "date-time"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "DateTimeNullableFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateTimeNullableFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateTimeNullableFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
...
      "DecimalFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecimalFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DecimalFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
...
      "GetAllToDoMasterReqDto": {
        "type": "object",
        "properties": {
          "toDoTaskId": {
            "$ref": "#/components/schemas/Int32FilterExpression"
          },
          "title": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "description": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "isComplete": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "repeatFrequency": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "createdDate": {
            "$ref": "#/components/schemas/DateTimeFilterExpression"
          },
          "dueDate": {
            "$ref": "#/components/schemas/DateTimeNullableFilterExpression"
          },
          "updatedDate": {
            "$ref": "#/components/schemas/DateTimeNullableFilterExpression"
          },
          "completedDate": {
            "$ref": "#/components/schemas/DateTimeNullableFilterExpression"
          },
          "createdById": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "assignedToId": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "isAssignedToTeam": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "hasChecklist": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "hasReminder": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "percentageCompleted": {
            "$ref": "#/components/schemas/DecimalFilterExpression"
          },
          "isDeleted": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "isStarred": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "mediaAttachmentType": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "mediaAttachmentURL": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "options": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "additionalProperties": false
      },
...
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "Int32FilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Int32FilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Int32FilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "LogicalOperator": {
        "enum": [
          "And",
          "Or"
        ],
        "type": "string"
      },
      "PageDetails": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sortBy": {
            "type": "string",
            "nullable": true
          },
          "sortDescending": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ResponseMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
...
      "ToDoChecklistResDto": {
        "type": "object",
        "properties": {
          "toDoChecklistId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ToDoMasterResDetailDto": {
        "type": "object",
        "properties": {
          "toDoChecklistToDoTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToDoChecklistResDto"
            },
            "nullable": true
          },
          "toDoReminderToDoTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToDoReminderResDto"
            },
            "nullable": true
          },
          "toDoMasterCreatedByIdfk": {
            "$ref": "#/components/schemas/AspNetUserResDto"
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "repeatFrequency": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "isAssignedToTeam": {
            "type": "boolean"
          },
          "hasChecklist": {
            "type": "boolean"
          },
          "hasReminder": {
            "type": "boolean"
          },
          "percentageCompleted": {
            "type": "number",
            "format": "double"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "mediaAttachmentType": {
            "type": "string",
            "nullable": true
          },
          "mediaAttachmentURL": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoMasterResDetailDtoSingleResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToDoMasterResDetailDto"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoMasterResDto": {
        "type": "object",
        "properties": {
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "repeatFrequency": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "isAssignedToTeam": {
            "type": "boolean"
          },
          "hasChecklist": {
            "type": "boolean"
          },
          "hasReminder": {
            "type": "boolean"
          },
          "percentageCompleted": {
            "type": "number",
            "format": "double"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "mediaAttachmentType": {
            "type": "string",
            "nullable": true
          },
          "mediaAttachmentURL": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoMasterResDtoListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToDoMasterResDto"
            },
            "nullable": true
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageDetails"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoMasterResDtoSingleResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToDoMasterResDto"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ToDoReminderResDto": {
        "type": "object",
        "properties": {
          "reminderId": {
            "type": "integer",
            "format": "int32"
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "setById": {
            "type": "string",
            "nullable": true
          },
          "setForDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "sentAtDateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateToDoMasterReqDto": {
        "type": "object",
        "properties": {
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "repeatFrequency": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "isAssignedToTeam": {
            "type": "boolean"
          },
          "hasChecklist": {
            "type": "boolean"
          },
          "hasReminder": {
            "type": "boolean"
          },
          "percentageCompleted": {
            "type": "number",
            "format": "double"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "mediaAttachmentType": {
            "type": "string",
            "nullable": true
          },
          "mediaAttachmentURL": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}